+++ index.d.ts
@@ -1,8 +1,9 @@
 import { FastifyPluginAsync } from 'fastify'
 
 export interface MercuriusExplainOptions {
-  enabled: boolean | (()=> boolean)
+  enabled: boolean | (() => boolean)
   gateway?: boolean
+  federated?: boolean
 }
 
 declare const mercuriusExplain: FastifyPluginAsync<MercuriusExplainOptions>
