+++ lib/constant.js
@@ -1,4 +1,9 @@
 export const DEFAULT_OPTIONS = {
   enabled: true,
-  gateway: false
+  gateway: false,
+  federated: false
+}
+
+export const FEDERATION_HEADER = {
+  'x-mercurius-explain-federation': true
 }
